requirements.txt
annotated-types>=0.7.0
anyio>=4.9.0
cachetools>=5.5.2
certifi>=2025.1.31
charset-normalizer>=3.4.1
Cython>=3.1.1
exceptiongroup>=1.2.2
filelock>=3.18.0
fsspec>=2025.3.2
google-ai-generativelanguage>=0.6.15
google-api-core>=2.24.2
google-api-python-client>=2.165.0
google-auth>=2.38.0
google-auth-httplib2>=0.2.0
google-cloud>=0.34.0
google-cloud-speech>=2.31.1
google-cloud-vision>=3.10.1
google-genai>=1.7.0
google-generativeai>=0.8.4
googleapis-common-protos>=1.69.2
grpcio>=1.71.0
grpcio-status>=1.71.0
h11>=0.14.0
h2>=4.2.0
hf-xet>=1.1.0
hpack>=4.1.0
httpcore>=1.0.7
httplib2>=0.22.0
httpx>=0.28.1
huggingface-hub>=0.31.1
hyperframe>=6.1.0
idna>=3.10
Jinja2>=3.1.6
joblib>=1.5.0
MarkupSafe>=3.0.2
mpmath>=1.3.0
networkx>=3.2.1
numpy>=2.0.2
packaging>=25.0
pillow>=11.2.1
portalocker>=2.10.1
proto-plus>=1.26.1
protobuf>=4.25.3
pyasn1>=0.6.1
pyasn1_modules>=0.4.1
pydantic>=2.10.6
pydantic_core>=2.27.2
pyparsing>=3.2.2
PyYAML>=6.0.2
qdrant-client>=1.14.2
regex>=2024.11.6
requests>=2.32.3
rsa>=4.9
safetensors>=0.5.3
scikit-learn>=1.6.1
scipy>=1.13.1
sentence-transformers>=4.1.0
sniffio>=1.3.1
sympy>=1.14.0
threadpoolctl>=3.6.0
tokenizers>=0.21.1
torch>=2.7.0
tqdm>=4.67.1
transformers>=4.51.3
typing_extensions>=4.12.2
uritemplate>=4.1.1
urllib3>=2.3.0
websockets>=15.0.1
# New Dependencies
fastapi
uvicorn
celery[redis]
redis
python-multipart
# apscheduler (already used in scheduler.py)
apscheduler
python-dotenv
sqlalchemy
passlib[bcrypt]
pyjwt
argon2-cffi
psutil
kss
prometheus-fastapi-instrumentator
